WIP: general codeinspection

+++ stealth/runmode/setmode.cc
@@ -2,16 +2,7 @@
 
 void RunMode::setMode(Mode mode)
 {
-    switch ((d_mode = mode))
-    {
-        default:
-        break;
-
-        case LEAVE:
-//            Lock::unlockRunFile();  // unlock the run file
-//            Lock::unlinkRunFile();  // remove the run file
-            unlink(Options::instance().runFile().c_str());
-        break;
-    }
+    if ((d_mode = mode) == LEAVE)
+        unlink(Options::instance().runFile().c_str());
 }
